trainer.py
#!/usr/bin/env python3
#
import os, sys
import pdb
from torch import nn
from transformers import Trainer


class SemaTrainer(Trainer):
    """docstring for SemaTrainer"""
    def compute_loss(self, model, inputs, return_outputs=False):
        labels = inputs.get("labels")
        # forward pass
        outputs = model(**inputs)
        logits = outputs.get("logits")
        # compute custom loss (suppose one has 3 labels with different weights)
        loss_fct = nn.CrossEntropyLoss()
        loss = loss_fct(logits.view(-1, self.model.config.vocab_size), labels.view(-1))
        return (loss, outputs) if return_outputs else loss